refactor: CarService의 우승자 찾는 기능 책임 분리
